refactor, pass tests
